Move all processors to their own module
