+++ src/engine/constants.rs
@@ -1,0 +1,3 @@
+pub const MERCHANT: &str = "merchant";
+pub const PDA_SEED: &[u8] = b"sol_payment_processor";
+pub const PROGRAM_OWNER: &str = "44fVncfVm5fB8VsRBwVZW75FdR1nSVUKcf9nUa4ky6qN";
