+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod instruction;
 pub mod processor;
 pub mod state;
 pub mod utils;
+pub mod engine;
